fix -wNarrowing warning

+++ 6364 Worlds/src/main/include/commands/IndexCommands/ShootLow.hpp
@@ -40,7 +40,7 @@ namespace IndexCommands{
     private:
       IndexSubsystem* subsystem;
       bool isFinished;
-      double startTime;
+      int64_t startTime;
 
   };
 }
